Moved Slide Animation to another file
