+++ libgloss/lib/config/routes.dart
@@ -33,31 +33,3 @@ class LibglossRoutes {
     return HomeNew();
   }
 }
-
-class SlideRoute extends PageRouteBuilder {
-  final Widget page;
-  final RouteSettings settings;
-  SlideRoute({required this.page, required this.settings})
-      : super(
-          pageBuilder: (
-            BuildContext context,
-            Animation<double> animation,
-            Animation<double> secondaryAnimation,
-          ) =>
-              page,
-          settings: settings,
-          transitionsBuilder: (
-            BuildContext context,
-            Animation<double> animation,
-            Animation<double> secondaryAnimation,
-            Widget child,
-          ) =>
-              SlideTransition(
-            position: Tween<Offset>(
-              begin: const Offset(1, 0),
-              end: Offset.zero,
-            ).animate(animation),
-            child: child,
-          ),
-        );
-}
