+++ libgloss/lib/widgets/slide_route.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class SlideRoute extends PageRouteBuilder {
+  final Widget page;
+  final RouteSettings settings;
+  SlideRoute({
+    required this.page,
+    required this.settings,
+  }) : super(
+          pageBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+          ) =>
+              page,
+          settings: settings,
+          transitionsBuilder: (
+            BuildContext context,
+            Animation<double> animation,
+            Animation<double> secondaryAnimation,
+            Widget child,
+          ) =>
+              SlideTransition(
+            position: Tween<Offset>(
+              begin: const Offset(1, 0),
+              end: Offset.zero,
+            ).animate(
+              animation,
+            ),
+            child: child,
+          ),
+        );
+}
